Change vertices to range 0 ~ size (was -1.0 ~ 1.0)

+++ population.h
@@ -73,10 +73,10 @@ private:
 	
 	// GENES
 	Scalar** colors;
-	Point2f*** solutions;
+	Point2i*** solutions;
 	bool* selected;
 	Scalar** p_colors;
-	Point2f*** p_solutions;
+	Point2i*** p_solutions;
 	
 	int parentsAmount, childsAmount;
